modified:   get_next_line.c 	modified:   get_next_line.h 	modified:   get_next_line_utils.c

+++ get_next_line.h
@@ -6,7 +6,7 @@
 /*   By: btan <btan@student.42singapore.sg>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/01 15:24:48 by btan              #+#    #+#             */
-/*   Updated: 2023/10/16 14:46:56 by btan             ###   ########.fr       */
+/*   Updated: 2023/10/17 12:35:06 by btan             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,7 +19,7 @@
 # define BUFFER_SIZE	1
 # endif
 
-size_t	ft_strlen(const char *s);
+int	ft_strlen(const char *s);
 void	ft_bzero(void *s, size_t n);
 void	*ft_calloc(size_t nmemb, size_t size);
 int	is_newline(const char *s);
